+++ task/lib/task.dart
@@ -1,3 +1,4 @@
+export 'package:task/src/functions/set_config.dart';
 export 'package:task/src/functions/show_exception_dialog.dart';
 export 'package:task/src/routes/configure_taskserver_route.dart';
 export 'package:task/src/routes/detail_route.dart';